Conflicts:
	.gitignore
	src/Broadway/Bundle/BroadwayBundle/Resources/config/services.xml

+++ src/Broadway/Bundle/BroadwayBundle/Command/SchemaEventStoreDropCommand.php
@@ -11,23 +11,22 @@
 
 namespace Broadway\Bundle\BroadwayBundle\Command;
 
-use Broadway\EventStore\DBALEventStore;
-use Doctrine\Bundle\DoctrineBundle\Command\DoctrineCommand;
 use Exception;
-use RuntimeException;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
 /**
  * Drops the event store schema.
  */
-class SchemaEventStoreDropCommand extends DoctrineCommand
+class SchemaEventStoreDropCommand extends AbstractSchemaEventStoreCommand
 {
     /**
      * {@inheritDoc}
      */
     protected function configure()
     {
+        parent::configure();
+
         $this
             ->setName('broadway:event-store:schema:drop')
             ->setDescription('Drops the event store schema')
@@ -46,17 +45,26 @@ EOT
      */
     protected function execute(InputInterface $input, OutputInterface $output)
     {
-        $connection = $this->getDoctrineConnection('default');
+        if (!$this->connection) {
+            $output->writeln('<error>Could not drop Broadway event-store schema</error>');
+            $output->writeln(sprintf('<error>%s</error>', $this->exception->getMessage()));
+
+            return 1;
+        }
 
         $error = false;
         try {
-            $schemaManager = $connection->getSchemaManager();
+            $schemaManager = $this->connection->getSchemaManager();
             $eventStore    = $this->getEventStore();
 
             $table = $eventStore->configureTable();
-            $schemaManager->dropTable($table->getName());
+            if ($schemaManager->tablesExist(array($table->getName()))) {
+                $schemaManager->dropTable($table->getName());
+                $output->writeln('<info>Dropped Broadway event-store schema</info>');
+            } else {
+                $output->writeln('<info>Broadway event-store schema does not exist</info>');
+            }
 
-            $output->writeln('<info>Dropped Broadway event-store schema</info>');
         } catch (Exception $e) {
             $output->writeln('<error>Could not drop Broadway event-store schema</error>');
             $output->writeln(sprintf('<error>%s</error>', $e->getMessage()));
@@ -64,16 +72,5 @@ EOT
         }
 
         return $error ? 1 : 0;
-    }
-
-    private function getEventStore()
-    {
-        $eventStore = $this->getContainer()->get('broadway.event_store');
-
-        if (! $eventStore instanceof DBALEventStore) {
-            throw new RuntimeException("'broadway.event_store' must be configured as an instance of DBALEventStore");
-        }
-
-        return $eventStore;
     }
 }
